+++ @navikt/core/react/src/data/token-filter/TokenFilter.tsx
@@ -8,7 +8,7 @@ import type {
   QueryFilteringOptions,
   QueryFilteringProperties,
 } from "./TokenFilter.types";
-import { generateAutoCompleteOptions } from "./helpers/generate-autocomplete-options-v2";
+import { generateAutoCompleteOptions } from "./helpers/generate-autocomplete-options";
 import { parseQueryText } from "./helpers/parse-query-text";
 
 type TokenFilterProps = {
